requirements.txt
asttokens==2.4.1
biopython==1.83
bokeh==3.3.4
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dask==2024.1.1
debugpy==1.8.0
decorator==5.1.1
distributed==2024.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.47.2
fsspec==2023.12.2
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.28.0
ipynb==0.5.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
locket==1.0.0
lz4==4.3.3
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
seaborn==0.13.1
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.0.0
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
xyzservices==2023.10.1
zict==3.0.0
zipp==3.17.0
